Remove match controller, routes, and service; update organization routes in index
