--- backend/routes/matchRoutes.js
@@ -1,6 +1,0 @@
-const express = require('express');
-const matchController = require('../controllers/matchController');
-const router = express.Router();
-
-
-module.exports = router;
